split large classes
