+++ docs/articles/README.md
@@ -2,7 +2,7 @@
 
 Long-form articles introducing Kotlin architecture testing and Konture.
 
-1. [Architecture Drift in Android & KMP Monorepos: How Big Codebases Rot (and How to Stop It)](architecture-drift-android-kmp-monorepos.md)
+1. [Architecture Drift in Android & KMP Monorepos: How Big Codebases Rot and How to Stop It](architecture-drift-android-kmp-monorepos.md)
 2. [How to Ban a Class or Method in Code (And Why You Should)](how-to-ban-a-class-or-method.md)
 3. [Kotlin Architecture Tests: What They Are and Why They Matter](kotlin-architecture-tests-what-and-why.md)
 4. [Kotlin Architecture Tests: Why Konture Exists](kotlin-architecture-tests-why-konture-exists.md)
