+++ library/src/test/kotlin/io/github/baole/konture/ClassesShouldAssertionsTest.kt
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-@file:Suppress("LongMethod")
-
 package io.github.baole.konture
 
 import org.junit.jupiter.api.Assertions.assertEquals
@@ -13,8 +11,10 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 
 internal class ClassesShouldAssertionsTest : KontureScopeTestFixture() {
+    private fun builder() = ClassesRuleBuilder(projectGraph)
+
     @Test
-    fun `test classes should remaining assertions`() {
+    fun `test classes assertions`() {
         val testClass =
             ClassDeclaration(
                 name = "TestClass",
@@ -22,303 +22,13 @@ internal class ClassesShouldAssertionsTest : KontureScopeTestFixture() {
                 packageName = "com.test",
                 isInterface = false,
                 isAbstract = false,
-                annotations =
-                    listOf(
-                        AnnotationDeclaration(
-                            "MyAnno",
-                            "com.test.MyAnno",
-                            arguments =
-                                listOf(
-                                    AnnotationArgumentDeclaration("value", "foo"),
-                                    AnnotationArgumentDeclaration("num", "42"),
-                                ),
-                        ),
-                    ),
+                annotations = emptyList(),
                 imports = emptyList(),
                 referencedTypes = emptySet(),
                 filePath = "/src/TestClass.kt",
-                modifiers = setOf(Modifier.DATA, Modifier.INLINE),
-                visibility = Visibility.PROTECTED,
-                supertypes = listOf("com.test.SuperInterface", "com.test.SuperClass"),
-                kdocText = "Hello KDoc",
+                supertypes = listOf("SuperInterface", "SuperClass"),
+                kdocText = "Some documentation",
             )
-
-        fun builder() = ClassesRuleBuilder(projectGraph)
-
-        val assertSatisfyPred = builder().should().satisfy { it.name == "TestClass" }.getShouldAssertion()!!
-        val vSat1 = mutableListOf<String>()
-        assertSatisfyPred(testClass, emptyList(), vSat1)
-        assertTrue(vSat1.isEmpty())
-
-        val assertSatisfyPredFail = builder().should().satisfy { it.name == "Wrong" }.getShouldAssertion()!!
-        val vSat2 = mutableListOf<String>()
-        assertSatisfyPredFail(testClass, emptyList(), vSat2)
-        assertEquals(1, vSat2.size)
-        assertTrue(vSat2[0].contains("should satisfy: custom condition"))
-
-        val assertSatisfyDesc = builder().should().satisfy("be awesome").getShouldAssertion()!!
-        val vSat3 = mutableListOf<String>()
-        assertSatisfyDesc(testClass, emptyList(), vSat3)
-        assertEquals(1, vSat3.size)
-        assertTrue(vSat3[0].contains("should satisfy: be awesome"))
-
-        val assertCustomViolations =
-            builder()
-                .should()
-                .satisfy { cls, violations ->
-                    if (cls.name != "Different") {
-                        violations.add("custom failure")
-                    }
-                }.getShouldAssertion()!!
-        val vSat4 = mutableListOf<String>()
-        assertCustomViolations(testClass, emptyList(), vSat4)
-        assertEquals(1, vSat4.size)
-        assertEquals("custom failure", vSat4[0])
-
-        val assertStart = builder().should().haveNameStartingWith("Test").getShouldAssertion()!!
-        val vSt1 = mutableListOf<String>()
-        assertStart(testClass, emptyList(), vSt1)
-        assertTrue(vSt1.isEmpty())
-
-        val assertStartFail = builder().should().haveNameStartingWith("Wrong").getShouldAssertion()!!
-        val vSt2 = mutableListOf<String>()
-        assertStartFail(testClass, emptyList(), vSt2)
-        assertEquals(1, vSt2.size)
-
-        val assertEnd = builder().should().haveNameEndingWith("Class").getShouldAssertion()!!
-        val vEd1 = mutableListOf<String>()
-        assertEnd(testClass, emptyList(), vEd1)
-        assertTrue(vEd1.isEmpty())
-
-        val assertEndFail = builder().should().haveNameEndingWith("Wrong").getShouldAssertion()!!
-        val vEd2 = mutableListOf<String>()
-        assertEndFail(testClass, emptyList(), vEd2)
-        assertEquals(1, vEd2.size)
-
-        val assertPkgPred = builder().should().resideInAPackage { it.startsWith("com.t") }.getShouldAssertion()!!
-        val vPk1 = mutableListOf<String>()
-        assertPkgPred(testClass, emptyList(), vPk1)
-        assertTrue(vPk1.isEmpty())
-
-        val assertPkgPredFail =
-            builder()
-                .should()
-                .resideInAPackage {
-                    it.startsWith(
-                        "com.other",
-                    )
-                }.getShouldAssertion()!!
-        val vPk2 = mutableListOf<String>()
-        assertPkgPredFail(testClass, emptyList(), vPk2)
-        assertEquals(1, vPk2.size)
-
-        val assertNamePred = builder().should().haveName { it.length == 9 }.getShouldAssertion()!!
-        val vNm1 = mutableListOf<String>()
-        assertNamePred(testClass, emptyList(), vNm1)
-        assertTrue(vNm1.isEmpty())
-
-        val assertNamePredFail = builder().should().haveName("be 5 chars") { it.length == 5 }.getShouldAssertion()!!
-        val vNm2 = mutableListOf<String>()
-        assertNamePredFail(testClass, emptyList(), vNm2)
-        assertEquals(1, vNm2.size)
-        assertTrue(vNm2[0].contains("should have name matching: be 5 chars"))
-
-        val assertAllAnnosList = builder().should().haveAllAnnotationsOf(listOf("MyAnno")).getShouldAssertion()!!
-        val vAn1 = mutableListOf<String>()
-        assertAllAnnosList(testClass, emptyList(), vAn1)
-        assertTrue(vAn1.isEmpty())
-
-        val assertAllAnnosListFail =
-            builder()
-                .should()
-                .haveAllAnnotationsOf(
-                    listOf("MyAnno", "OtherAnno"),
-                ).getShouldAssertion()!!
-        val vAn2 = mutableListOf<String>()
-        assertAllAnnosListFail(testClass, emptyList(), vAn2)
-        assertEquals(1, vAn2.size)
-
-        val assertAnyAnnosList =
-            builder()
-                .should()
-                .haveAnyAnnotationOf(
-                    listOf("MyAnno", "OtherAnno"),
-                ).getShouldAssertion()!!
-        val vAn3 = mutableListOf<String>()
-        assertAnyAnnosList(testClass, emptyList(), vAn3)
-        assertTrue(vAn3.isEmpty())
-
-        val assertAnyAnnosVararg = builder().should().haveAnyAnnotationOf("Other1", "Other2").getShouldAssertion()!!
-        val vAn4 = mutableListOf<String>()
-        assertAnyAnnosVararg(testClass, emptyList(), vAn4)
-        assertEquals(1, vAn4.size)
-
-        val assertAnyAnnosListFail = builder().should().haveAnyAnnotationOf(listOf("OtherAnno")).getShouldAssertion()!!
-        val vAn5 = mutableListOf<String>()
-        assertAnyAnnosListFail(testClass, emptyList(), vAn5)
-        assertEquals(1, vAn5.size)
-
-        val assertArg1 = builder().should().haveAnnotationWithArgument("MyAnno", "value", "foo").getShouldAssertion()!!
-        val vArg1 = mutableListOf<String>()
-        assertArg1(testClass, emptyList(), vArg1)
-        assertTrue(vArg1.isEmpty())
-
-        val assertArg2 = builder().should().haveAnnotationWithArgument("MyAnno", "num", "42").getShouldAssertion()!!
-        val vArg2 = mutableListOf<String>()
-        assertArg2(testClass, emptyList(), vArg2)
-        assertTrue(vArg2.isEmpty())
-
-        val assertArgFailVal =
-            builder()
-                .should()
-                .haveAnnotationWithArgument(
-                    "MyAnno",
-                    "value",
-                    "bar",
-                ).getShouldAssertion()!!
-        val vArg3 = mutableListOf<String>()
-        assertArgFailVal(testClass, emptyList(), vArg3)
-        assertEquals(1, vArg3.size)
-
-        val assertArgFailName =
-            builder()
-                .should()
-                .haveAnnotationWithArgument(
-                    "MyAnno",
-                    "wrongArg",
-                    "foo",
-                ).getShouldAssertion()!!
-        val vArg4 = mutableListOf<String>()
-        assertArgFailName(testClass, emptyList(), vArg4)
-        assertEquals(1, vArg4.size)
-
-        val assertArgFailAnno =
-            builder()
-                .should()
-                .haveAnnotationWithArgument(
-                    "WrongAnno",
-                    "value",
-                    "foo",
-                ).getShouldAssertion()!!
-        val vArg5 = mutableListOf<String>()
-        assertArgFailAnno(testClass, emptyList(), vArg5)
-        assertEquals(1, vArg5.size)
-
-        val assertInterfaces = builder().should().beInterfaces().getShouldAssertion()!!
-        val vBi = mutableListOf<String>()
-        assertInterfaces(testClass, emptyList(), vBi)
-        assertEquals(1, vBi.size)
-
-        val assertEnums = builder().should().beEnums().getShouldAssertion()!!
-        val vBe = mutableListOf<String>()
-        assertEnums(testClass, emptyList(), vBe)
-        assertEquals(1, vBe.size)
-
-        val assertAbstract = builder().should().beAbstract().getShouldAssertion()!!
-        val vBab = mutableListOf<String>()
-        assertAbstract(testClass, emptyList(), vBab)
-        assertEquals(1, vBab.size)
-
-        val assertSealed = builder().should().beSealed().getShouldAssertion()!!
-        val vBse = mutableListOf<String>()
-        assertSealed(testClass, emptyList(), vBse)
-        assertEquals(1, vBse.size)
-
-        val assertData = builder().should().beData().getShouldAssertion()!!
-        val vBd = mutableListOf<String>()
-        assertData(testClass, emptyList(), vBd)
-        assertTrue(vBd.isEmpty())
-
-        val assertInline = builder().should().beInline().getShouldAssertion()!!
-        val vBin = mutableListOf<String>()
-        assertInline(testClass, emptyList(), vBin)
-        assertTrue(vBin.isEmpty())
-
-        val assertMod1 = builder().should().haveModifier(Modifier.DATA).getShouldAssertion()!!
-        val vMod1 = mutableListOf<String>()
-        assertMod1(testClass, emptyList(), vMod1)
-        assertTrue(vMod1.isEmpty())
-
-        val assertMod1Fail = builder().should().haveModifier(Modifier.ABSTRACT).getShouldAssertion()!!
-        val vMod2 = mutableListOf<String>()
-        assertMod1Fail(testClass, emptyList(), vMod2)
-        assertEquals(1, vMod2.size)
-
-        val assertAllModsList =
-            builder()
-                .should()
-                .haveAllModifiers(
-                    listOf(Modifier.DATA, Modifier.INLINE),
-                ).getShouldAssertion()!!
-        val vMod3 = mutableListOf<String>()
-        assertAllModsList(testClass, emptyList(), vMod3)
-        assertTrue(vMod3.isEmpty())
-
-        val assertAllModsListFail =
-            builder()
-                .should()
-                .haveAllModifiers(
-                    listOf(Modifier.DATA, Modifier.ABSTRACT),
-                ).getShouldAssertion()!!
-        val vMod4 = mutableListOf<String>()
-        assertAllModsListFail(testClass, emptyList(), vMod4)
-        assertEquals(1, vMod4.size)
-
-        val assertAnyModsList =
-            builder()
-                .should()
-                .haveAnyModifier(
-                    listOf(Modifier.ABSTRACT, Modifier.INLINE),
-                ).getShouldAssertion()!!
-        val vMod5 = mutableListOf<String>()
-        assertAnyModsList(testClass, emptyList(), vMod5)
-        assertTrue(vMod5.isEmpty())
-
-        val assertAnyModsListFail =
-            builder()
-                .should()
-                .haveAnyModifier(
-                    listOf(Modifier.ABSTRACT, Modifier.SEALED),
-                ).getShouldAssertion()!!
-        val vMod6 = mutableListOf<String>()
-        assertAnyModsListFail(testClass, emptyList(), vMod6)
-        assertEquals(1, vMod6.size)
-
-        val assertVis = builder().should().haveVisibility(Visibility.PROTECTED).getShouldAssertion()!!
-        val vVis1 = mutableListOf<String>()
-        assertVis(testClass, emptyList(), vVis1)
-        assertTrue(vVis1.isEmpty())
-
-        val assertVisFail = builder().should().haveVisibility(Visibility.PUBLIC).getShouldAssertion()!!
-        val vVis2 = mutableListOf<String>()
-        assertVisFail(testClass, emptyList(), vVis2)
-        assertEquals(1, vVis2.size)
-
-        val assertAnyVisList =
-            builder()
-                .should()
-                .haveAnyVisibility(
-                    listOf(Visibility.PROTECTED, Visibility.PUBLIC),
-                ).getShouldAssertion()!!
-        val vVis3 = mutableListOf<String>()
-        assertAnyVisList(testClass, emptyList(), vVis3)
-        assertTrue(vVis3.isEmpty())
-
-        val assertAnyVisListFail =
-            builder()
-                .should()
-                .haveAnyVisibility(
-                    listOf(Visibility.PRIVATE, Visibility.PUBLIC),
-                ).getShouldAssertion()!!
-        val vVis4 = mutableListOf<String>()
-        assertAnyVisListFail(testClass, emptyList(), vVis4)
-        assertEquals(1, vVis4.size)
-
-        assertTrue(builder().should().beProtected().getShouldAssertion() != null)
-        assertTrue(builder().should().bePublic().getShouldAssertion() != null)
-        assertTrue(builder().should().beInternal().getShouldAssertion() != null)
-        assertTrue(builder().should().bePrivate().getShouldAssertion() != null)
-
         val childClass =
             ClassDeclaration(
                 name = "ChildClass",
@@ -330,11 +40,254 @@ internal class ClassesShouldAssertionsTest : KontureScopeTestFixture() {
                 imports = emptyList(),
                 referencedTypes = emptySet(),
                 filePath = "/src/ChildClass.kt",
-                modifiers = emptySet(),
-                visibility = Visibility.PUBLIC,
-                supertypes = listOf("com.test.TestClass"),
-                kdocText = null,
+                supertypes = listOf("TestClass"),
             )
+
+        val assertResideSingle = builder().should().resideInAPackage("com.test").getShouldAssertion()!!
+        val v1 = mutableListOf<String>()
+        assertResideSingle(testClass, emptyList(), v1)
+        assertTrue(v1.isEmpty())
+
+        val assertResideList = builder().should().resideInAPackage(listOf("com.test")).getShouldAssertion()!!
+        val v2 = mutableListOf<String>()
+        assertResideList(testClass, emptyList(), v2)
+        assertTrue(v2.isEmpty())
+
+        val assertResideVararg = builder().should().resideInAPackage("com.test", "com.other").getShouldAssertion()!!
+        val v3 = mutableListOf<String>()
+        assertResideVararg(testClass, emptyList(), v3)
+        assertTrue(v3.isEmpty())
+
+        val assertNotResideSingle = builder().should().notResideInAPackage("com.other").getShouldAssertion()!!
+        val v4 = mutableListOf<String>()
+        assertNotResideSingle(testClass, emptyList(), v4)
+        assertTrue(v4.isEmpty())
+
+        val assertNotResideList = builder().should().notResideInAPackage(listOf("com.other")).getShouldAssertion()!!
+        val v5 = mutableListOf<String>()
+        assertNotResideList(testClass, emptyList(), v5)
+        assertTrue(v5.isEmpty())
+
+        val assertNotResideVararg =
+            builder().should().notResideInAPackage("com.other", "com.wrong").getShouldAssertion()!!
+        val v6 = mutableListOf<String>()
+        assertNotResideVararg(testClass, emptyList(), v6)
+        assertTrue(v6.isEmpty())
+
+        val assertModSingle = builder().should().resideInAModule(":app").getShouldAssertion()!!
+        val vMod1 = mutableListOf<String>()
+        assertModSingle(testClass, emptyList(), vMod1)
+        assertTrue(vMod1.isEmpty())
+
+        val assertModList = builder().should().resideInAModule(listOf(":app")).getShouldAssertion()!!
+        val vMod2 = mutableListOf<String>()
+        assertModList(testClass, emptyList(), vMod2)
+        assertTrue(vMod2.isEmpty())
+
+        val assertModVararg = builder().should().resideInAModule(":app", ":core").getShouldAssertion()!!
+        val vMod3 = mutableListOf<String>()
+        assertModVararg(testClass, emptyList(), vMod3)
+        assertTrue(vMod3.isEmpty())
+
+        val assertNotModSingle = builder().should().notResideInAModule(":forbidden").getShouldAssertion()!!
+        val vMod4 = mutableListOf<String>()
+        assertNotModSingle(testClass, emptyList(), vMod4)
+        assertTrue(vMod4.isEmpty())
+
+        val assertNotModList = builder().should().notResideInAModule(listOf(":forbidden")).getShouldAssertion()!!
+        val vMod5 = mutableListOf<String>()
+        assertNotModList(testClass, emptyList(), vMod5)
+        assertTrue(vMod5.isEmpty())
+
+        val assertNotModVararg =
+            builder().should().notResideInAModule(":forbidden", ":other").getShouldAssertion()!!
+        val vMod6 = mutableListOf<String>()
+        assertNotModVararg(testClass, emptyList(), vMod6)
+        assertTrue(vMod6.isEmpty())
+
+        val assertNameSingle = builder().should().haveName("TestClass").getShouldAssertion()!!
+        val vN1 = mutableListOf<String>()
+        assertNameSingle(testClass, emptyList(), vN1)
+        assertTrue(vN1.isEmpty())
+
+        val assertNameList = builder().should().haveName(listOf("TestClass")).getShouldAssertion()!!
+        val vN2 = mutableListOf<String>()
+        assertNameList(testClass, emptyList(), vN2)
+        assertTrue(vN2.isEmpty())
+
+        val assertNameVararg = builder().should().haveName("TestClass", "OtherClass").getShouldAssertion()!!
+        val vN3 = mutableListOf<String>()
+        assertNameVararg(testClass, emptyList(), vN3)
+        assertTrue(vN3.isEmpty())
+
+        val assertNotNameSingle = builder().should().notHaveName("WrongClass").getShouldAssertion()!!
+        val vN4 = mutableListOf<String>()
+        assertNotNameSingle(testClass, emptyList(), vN4)
+        assertTrue(vN4.isEmpty())
+
+        val assertNotNameList = builder().should().notHaveName(listOf("WrongClass")).getShouldAssertion()!!
+        val vN5 = mutableListOf<String>()
+        assertNotNameList(testClass, emptyList(), vN5)
+        assertTrue(vN5.isEmpty())
+
+        val assertNotNameVararg = builder().should().notHaveName("Wrong1", "Wrong2").getShouldAssertion()!!
+        val vN6 = mutableListOf<String>()
+        assertNotNameVararg(testClass, emptyList(), vN6)
+        assertTrue(vN6.isEmpty())
+
+        val assertStartSingle = builder().should().haveNameStartingWith("Test").getShouldAssertion()!!
+        val vS1 = mutableListOf<String>()
+        assertStartSingle(testClass, emptyList(), vS1)
+        assertTrue(vS1.isEmpty())
+
+        val assertStartList = builder().should().haveNameStartingWith(listOf("Test")).getShouldAssertion()!!
+        val vS2 = mutableListOf<String>()
+        assertStartList(testClass, emptyList(), vS2)
+        assertTrue(vS2.isEmpty())
+
+        val assertStartVararg = builder().should().haveNameStartingWith("Test", "Other").getShouldAssertion()!!
+        val vS3 = mutableListOf<String>()
+        assertStartVararg(testClass, emptyList(), vS3)
+        assertTrue(vS3.isEmpty())
+
+        val assertNotStartSingle = builder().should().notHaveNameStartingWith("Wrong").getShouldAssertion()!!
+        val vS4 = mutableListOf<String>()
+        assertNotStartSingle(testClass, emptyList(), vS4)
+        assertTrue(vS4.isEmpty())
+
+        val assertNotStartList = builder().should().notHaveNameStartingWith(listOf("Wrong")).getShouldAssertion()!!
+        val vS5 = mutableListOf<String>()
+        assertNotStartList(testClass, emptyList(), vS5)
+        assertTrue(vS5.isEmpty())
+
+        val assertNotStartVararg =
+            builder().should().notHaveNameStartingWith("Wrong1", "Wrong2").getShouldAssertion()!!
+        val vS6 = mutableListOf<String>()
+        assertNotStartVararg(testClass, emptyList(), vS6)
+        assertTrue(vS6.isEmpty())
+
+        val assertEndSingle = builder().should().haveNameEndingWith("Class").getShouldAssertion()!!
+        val vE1 = mutableListOf<String>()
+        assertEndSingle(testClass, emptyList(), vE1)
+        assertTrue(vE1.isEmpty())
+
+        val assertEndList = builder().should().haveNameEndingWith(listOf("Class")).getShouldAssertion()!!
+        val vE2 = mutableListOf<String>()
+        assertEndList(testClass, emptyList(), vE2)
+        assertTrue(vE2.isEmpty())
+
+        val assertEndVararg = builder().should().haveNameEndingWith("Class", "Other").getShouldAssertion()!!
+        val vE3 = mutableListOf<String>()
+        assertEndVararg(testClass, emptyList(), vE3)
+        assertTrue(vE3.isEmpty())
+
+        val assertNotEndSingle = builder().should().notHaveNameEndingWith("Wrong").getShouldAssertion()!!
+        val vE4 = mutableListOf<String>()
+        assertNotEndSingle(testClass, emptyList(), vE4)
+        assertTrue(vE4.isEmpty())
+
+        val assertNotEndList = builder().should().notHaveNameEndingWith(listOf("Wrong")).getShouldAssertion()!!
+        val vE5 = mutableListOf<String>()
+        assertNotEndList(testClass, emptyList(), vE5)
+        assertTrue(vE5.isEmpty())
+
+        val assertNotEndVararg = builder().should().notHaveNameEndingWith("Wrong1", "Wrong2").getShouldAssertion()!!
+        val vE6 = mutableListOf<String>()
+        assertNotEndVararg(testClass, emptyList(), vE6)
+        assertTrue(vE6.isEmpty())
+
+        val assertMatchSingle = builder().should().haveNameMatching("Test*").getShouldAssertion()!!
+        val vM1 = mutableListOf<String>()
+        assertMatchSingle(testClass, emptyList(), vM1)
+        assertTrue(vM1.isEmpty())
+
+        val assertMatchList = builder().should().haveNameMatching(listOf("Test*")).getShouldAssertion()!!
+        val vM2 = mutableListOf<String>()
+        assertMatchList(testClass, emptyList(), vM2)
+        assertTrue(vM2.isEmpty())
+
+        val assertMatchVararg = builder().should().haveNameMatching("Test*", "Other*").getShouldAssertion()!!
+        val vM3 = mutableListOf<String>()
+        assertMatchVararg(testClass, emptyList(), vM3)
+        assertTrue(vM3.isEmpty())
+
+        val assertNotMatchSingle = builder().should().notHaveNameMatching("Wrong*").getShouldAssertion()!!
+        val vM4 = mutableListOf<String>()
+        assertNotMatchSingle(testClass, emptyList(), vM4)
+        assertTrue(vM4.isEmpty())
+
+        val assertNotMatchList = builder().should().notHaveNameMatching(listOf("Wrong*")).getShouldAssertion()!!
+        val vM5 = mutableListOf<String>()
+        assertNotMatchList(testClass, emptyList(), vM5)
+        assertTrue(vM5.isEmpty())
+
+        val assertNotMatchVararg = builder().should().notHaveNameMatching("Wrong1*", "Wrong2*").getShouldAssertion()!!
+        val vM6 = mutableListOf<String>()
+        assertNotMatchVararg(testClass, emptyList(), vM6)
+        assertTrue(vM6.isEmpty())
+
+        val assertAnnotSingle = builder().should().haveAnnotationOf("MyAnnotation").getShouldAssertion()!!
+        val annotatedClass =
+            testClass.copy(annotations = listOf(AnnotationDeclaration("MyAnnotation", "com.test.MyAnnotation")))
+        val vA1 = mutableListOf<String>()
+        assertAnnotSingle(annotatedClass, emptyList(), vA1)
+        assertTrue(vA1.isEmpty())
+
+        val assertAnnotList = builder().should().haveAnnotationOf(listOf("MyAnnotation")).getShouldAssertion()!!
+        val vA2 = mutableListOf<String>()
+        assertAnnotList(annotatedClass, emptyList(), vA2)
+        assertTrue(vA2.isEmpty())
+
+        val assertAnnotVararg = builder().should().haveAnnotationOf("MyAnnotation", "Other").getShouldAssertion()!!
+        val vA3 = mutableListOf<String>()
+        assertAnnotVararg(annotatedClass, emptyList(), vA3)
+        assertTrue(vA3.isEmpty())
+
+        val assertNotAnnotSingle = builder().should().notHaveAnnotationOf("WrongAnnotation").getShouldAssertion()!!
+        val vA4 = mutableListOf<String>()
+        assertNotAnnotSingle(annotatedClass, emptyList(), vA4)
+        assertTrue(vA4.isEmpty())
+
+        val assertNotAnnotList = builder().should().notHaveAnnotationOf(listOf("WrongAnnotation")).getShouldAssertion()!!
+        val vA5 = mutableListOf<String>()
+        assertNotAnnotList(annotatedClass, emptyList(), vA5)
+        assertTrue(vA5.isEmpty())
+
+        val assertNotAnnotVararg =
+            builder().should().notHaveAnnotationOf("Wrong1", "Wrong2").getShouldAssertion()!!
+        val vA6 = mutableListOf<String>()
+        assertNotAnnotVararg(annotatedClass, emptyList(), vA6)
+        assertTrue(vA6.isEmpty())
+
+        val assertAllAnnotList = builder().should().haveAllAnnotationsOf(listOf("MyAnnotation")).getShouldAssertion()!!
+        val vA7 = mutableListOf<String>()
+        assertAllAnnotList(annotatedClass, emptyList(), vA7)
+        assertTrue(vA7.isEmpty())
+
+        val assertAllAnnotVararg = builder().should().haveAllAnnotationsOf("MyAnnotation").getShouldAssertion()!!
+        val vA8 = mutableListOf<String>()
+        assertAllAnnotVararg(annotatedClass, emptyList(), vA8)
+        assertTrue(vA8.isEmpty())
+
+        val assertAnyAnnotList = builder().should().haveAnyAnnotationOf(listOf("MyAnnotation", "Wrong")).getShouldAssertion()!!
+        val vA9 = mutableListOf<String>()
+        assertAnyAnnotList(annotatedClass, emptyList(), vA9)
+        assertTrue(vA9.isEmpty())
+
+        val assertAnyAnnotVararg = builder().should().haveAnyAnnotationOf("MyAnnotation", "Wrong").getShouldAssertion()!!
+        val vA10 = mutableListOf<String>()
+        assertAnyAnnotVararg(annotatedClass, emptyList(), vA10)
+        assertTrue(vA10.isEmpty())
+
+        val assertChildSingle = builder().should().beChildOf("com.test.TestClass").getShouldAssertion()!!
+        val vC1 = mutableListOf<String>()
+        assertChildSingle(childClass, listOf(testClass, childClass), vC1)
+        assertTrue(vC1.isEmpty())
+
+        val assertChildFail = builder().should().beChildOf("com.test.WrongClass").getShouldAssertion()!!
+        val vC2 = mutableListOf<String>()
+        assertChildFail(childClass, listOf(testClass, childClass), vC2)
+        assertEquals(1, vC2.size)
 
         val assertAssignableFrom = builder().should().beAssignableFrom("com.test.ChildClass").getShouldAssertion()!!
         val vAssign1 = mutableListOf<String>()
@@ -350,375 +303,5 @@ internal class ClassesShouldAssertionsTest : KontureScopeTestFixture() {
         val vAssign3 = mutableListOf<String>()
         assertAssignableFromSelf(testClass, listOf(testClass, childClass), vAssign3)
         assertTrue(vAssign3.isEmpty())
-
-        val assertAss = builder().should().beAssignableTo("com.test.SuperInterface").getShouldAssertion()!!
-        val vAss1 = mutableListOf<String>()
-        assertAss(testClass, emptyList(), vAss1)
-        assertTrue(vAss1.isEmpty())
-
-        val assertAssFail = builder().should().beAssignableTo("com.test.WrongSuper").getShouldAssertion()!!
-        val vAss2 = mutableListOf<String>()
-        assertAssFail(testClass, emptyList(), vAss2)
-        assertEquals(1, vAss2.size)
-
-        val assertAssAnyList =
-            builder()
-                .should()
-                .beAssignableToAnyOf(
-                    listOf("com.test.SuperInterface", "com.test.Wrong"),
-                ).getShouldAssertion()!!
-        val vAss3 = mutableListOf<String>()
-        assertAssAnyList(testClass, emptyList(), vAss3)
-        assertTrue(vAss3.isEmpty())
-
-        val assertAssAnyVararg =
-            builder()
-                .should()
-                .beAssignableToAnyOf(
-                    "com.test.Wrong",
-                    "com.test.Wrong2",
-                ).getShouldAssertion()!!
-        val vAss4 = mutableListOf<String>()
-        assertAssAnyVararg(testClass, emptyList(), vAss4)
-        assertEquals(1, vAss4.size)
-
-        val assertAssAllList =
-            builder()
-                .should()
-                .beAssignableToAllOf(
-                    listOf("com.test.SuperInterface", "com.test.SuperClass"),
-                ).getShouldAssertion()!!
-        val vAss5 = mutableListOf<String>()
-        assertAssAllList(testClass, emptyList(), vAss5)
-        assertTrue(vAss5.isEmpty())
-
-        val assertAssAllVarargFail =
-            builder()
-                .should()
-                .beAssignableToAllOf(
-                    "com.test.SuperInterface",
-                    "com.test.Wrong",
-                ).getShouldAssertion()!!
-        val vAss6 = mutableListOf<String>()
-        assertAssAllVarargFail(testClass, emptyList(), vAss6)
-        assertEquals(1, vAss6.size)
-
-        val grandParent =
-            ClassDeclaration(
-                name = "GrandParent",
-                fqName = "com.example.GrandParent",
-                packageName = "com.example",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = emptyList(),
-                referencedTypes = emptySet(),
-                filePath = "/src/GrandParent.kt",
-                supertypes = emptyList(),
-            )
-        val parent =
-            ClassDeclaration(
-                name = "Parent",
-                fqName = "com.example.Parent",
-                packageName = "com.example",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = emptyList(),
-                referencedTypes = emptySet(),
-                filePath = "/src/Parent.kt",
-                supertypes = listOf("GrandParent"),
-            )
-        val child =
-            ClassDeclaration(
-                name = "Child",
-                fqName = "com.example.Child",
-                packageName = "com.example",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = emptyList(),
-                referencedTypes = emptySet(),
-                filePath = "/src/Child.kt",
-                supertypes = listOf("Parent"),
-            )
-        val allHierarchy = listOf(grandParent, parent, child)
-        val assertAssTransitive = builder().should().beAssignableTo("GrandParent").getShouldAssertion()!!
-        val vAssTransitive = mutableListOf<String>()
-        assertAssTransitive(child, allHierarchy, vAssTransitive)
-        assertTrue(vAssTransitive.isEmpty())
-
-        val assertAssAnyTransitive =
-            builder().should().beAssignableToAnyOf(
-                "GrandParent",
-                "WrongType",
-            ).getShouldAssertion()!!
-        val vAssAnyTransitive = mutableListOf<String>()
-        assertAssAnyTransitive(child, allHierarchy, vAssAnyTransitive)
-        assertTrue(vAssAnyTransitive.isEmpty())
-
-        val assertAssAllTransitive =
-            builder().should().beAssignableToAllOf(
-                "GrandParent",
-                "Parent",
-            ).getShouldAssertion()!!
-        val vAssAllTransitive = mutableListOf<String>()
-        assertAssAllTransitive(child, allHierarchy, vAssAllTransitive)
-        assertTrue(vAssAllTransitive.isEmpty())
-
-        val assertKDoc = builder().should().beDocumentedWithKDoc().getShouldAssertion()!!
-        val vKd1 = mutableListOf<String>()
-        assertKDoc(testClass, emptyList(), vKd1)
-        assertTrue(vKd1.isEmpty())
-
-        val testClassNoKDoc = testClass.copy(kdocText = null)
-        val vKd2 = mutableListOf<String>()
-        assertKDoc(testClassNoKDoc, emptyList(), vKd2)
-        assertEquals(1, vKd2.size)
-
-        val testClassBlankKDoc = testClass.copy(kdocText = "   ")
-        val vKd3 = mutableListOf<String>()
-        assertKDoc(testClassBlankKDoc, emptyList(), vKd3)
-        assertEquals(1, vKd3.size)
-
-        val accessorClass =
-            ClassDeclaration(
-                name = "Accessor",
-                fqName = "com.other.Accessor",
-                packageName = "com.other",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = listOf("com.test.TestClass"),
-                referencedTypes = setOf("com.test.TestClass"),
-                filePath = "/src/Accessor.kt",
-            )
-        val assertAccess = builder().should().onlyBeAccessedByAnyPackage("com.other").getShouldAssertion()!!
-        val vAcc1 = mutableListOf<String>()
-        assertAccess(testClass, listOf(testClass, accessorClass), vAcc1)
-        assertTrue(vAcc1.isEmpty())
-
-        val assertAccessFail = builder().should().onlyBeAccessedByAnyPackage("com.allowed.*").getShouldAssertion()!!
-        val vAcc2 = mutableListOf<String>()
-        assertAccessFail(testClass, listOf(testClass, accessorClass), vAcc2)
-        assertEquals(1, vAcc2.size)
-
-        val assertDep = builder().should().onlyDependOnClassesInAnyPackage("com.other").getShouldAssertion()!!
-        val vDep1 = mutableListOf<String>()
-        assertDep(testClass, listOf(testClass, accessorClass), vDep1)
-        assertTrue(vDep1.isEmpty())
-
-        val dependentClass =
-            ClassDeclaration(
-                name = "Dependent",
-                fqName = "com.test.TestClass",
-                packageName = "com.test",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = listOf("com.other.Accessor"),
-                referencedTypes = setOf("com.other.Accessor"),
-                filePath = "/src/TestClass.kt",
-            )
-        val assertDepFail = builder().should().onlyDependOnClassesInAnyPackage("com.allowed").getShouldAssertion()!!
-        val vDep2 = mutableListOf<String>()
-        assertDepFail(dependentClass, listOf(dependentClass, accessorClass), vDep2)
-        assertEquals(1, vDep2.size)
-
-        val classWithStd =
-            ClassDeclaration(
-                name = "ClassWithStd",
-                fqName = "com.test.ClassWithStd",
-                packageName = "com.test",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = listOf("java.util.UUID", "kotlin.collections.List", "javax.inject.Inject"),
-                referencedTypes = emptySet(),
-                filePath = "/src/ClassWithStd.kt",
-            )
-        val assertDepStd = builder().should().onlyDependOnClassesInAnyPackage("com.test").getShouldAssertion()!!
-        val vDepStd = mutableListOf<String>()
-        assertDepStd(classWithStd, listOf(classWithStd), vDepStd)
-        assertTrue(vDepStd.isEmpty())
-
-        val classWithExt =
-            ClassDeclaration(
-                name = "ClassWithExt",
-                fqName = "com.test.ClassWithExt",
-                packageName = "com.test",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = listOf("org.json.JSONObject"),
-                referencedTypes = emptySet(),
-                filePath = "/src/ClassWithExt.kt",
-            )
-        val vDepExt = mutableListOf<String>()
-        assertDepStd(classWithExt, listOf(classWithExt), vDepExt)
-        assertEquals(1, vDepExt.size)
-
-        val assertNotDepExt = builder().should().notDependOnClassesInAnyPackage("org.json..").getShouldAssertion()!!
-        val vNotDepExt = mutableListOf<String>()
-        assertNotDepExt(classWithExt, listOf(classWithExt), vNotDepExt)
-        assertEquals(1, vNotDepExt.size)
-
-        val func1 =
-            FunctionDeclaration(
-                name = "myFunc",
-                visibility = Visibility.PUBLIC,
-                modifiers = emptySet(),
-                returnType = "Unit",
-                parameters = emptyList(),
-                annotations = emptyList(),
-                kdocText = null,
-                isExtension = false,
-            )
-        val classWithFunc = testClass.copy(functions = listOf(func1))
-        val assertFunc =
-            builder()
-                .should()
-                .allFunctions {
-                    bePublic()
-                    beInline()
-                }.getShouldAssertion()!!
-        val vFn = mutableListOf<String>()
-        assertFunc(classWithFunc, emptyList(), vFn)
-        assertEquals(1, vFn.size)
-        assertTrue(vFn[0].contains("Function myFunc in class com.test.TestClass has violations"))
-
-        val prop1 =
-            PropertyDeclaration(
-                name = "myProp",
-                visibility = Visibility.PRIVATE,
-                modifiers = emptySet(),
-                type = "String",
-                isVal = true,
-                annotations = emptyList(),
-                kdocText = null,
-                isExtension = false,
-            )
-        val classWithProp = testClass.copy(properties = listOf(prop1))
-        val assertProp =
-            builder()
-                .should()
-                .allProperties {
-                    bePublic()
-                }.getShouldAssertion()!!
-        val vPr = mutableListOf<String>()
-        assertProp(classWithProp, emptyList(), vPr)
-        assertEquals(1, vPr.size)
-        assertTrue(vPr[0].contains("Property myProp in class com.test.TestClass has violations"))
-    }
-
-    @Test
-    fun `test classes notBeAccessedByAnyPackage flags accessors in forbidden packages`() {
-        fun builder() = ClassesRuleBuilder(projectGraph)
-
-        val target =
-            ClassDeclaration(
-                name = "DomainModel",
-                fqName = "com.acme.domain.DomainModel",
-                packageName = "com.acme.domain",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = emptyList(),
-                referencedTypes = emptySet(),
-                filePath = "/src/DomainModel.kt",
-            )
-        val forbiddenAccessor =
-            ClassDeclaration(
-                name = "WebController",
-                fqName = "com.acme.web.WebController",
-                packageName = "com.acme.web",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = listOf("com.acme.domain.DomainModel"),
-                referencedTypes = setOf("com.acme.domain.DomainModel"),
-                filePath = "/src/WebController.kt",
-            )
-        val allowedAccessor =
-            ClassDeclaration(
-                name = "DomainService",
-                fqName = "com.acme.domain.DomainService",
-                packageName = "com.acme.domain",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = listOf("com.acme.domain.DomainModel"),
-                referencedTypes = setOf("com.acme.domain.DomainModel"),
-                filePath = "/src/DomainService.kt",
-            )
-
-        val assertForbidden = builder().should().notBeAccessedByAnyPackage("..web..").getShouldAssertion()!!
-        val vForbidden = mutableListOf<String>()
-        assertForbidden(target, listOf(target, forbiddenAccessor, allowedAccessor), vForbidden)
-        assertEquals(1, vForbidden.size)
-
-        val assertClean = builder().should().notBeAccessedByAnyPackage("..web..").getShouldAssertion()!!
-        val vClean = mutableListOf<String>()
-        assertClean(target, listOf(target, allowedAccessor), vClean)
-        assertTrue(vClean.isEmpty())
-
-        val assertVacuous = builder().should().notBeAccessedByAnyPackage("..web..").getShouldAssertion()!!
-        val vVacuous = mutableListOf<String>()
-        assertVacuous(target, listOf(target), vVacuous)
-        assertTrue(vVacuous.isEmpty())
-    }
-
-    @Test
-    fun `test signature leak in constructor parameters`() {
-        val entityAnnotation = AnnotationDeclaration("Entity", "jakarta.persistence.Entity")
-        val entityClass =
-            ClassDeclaration(
-                name = "UserEntity",
-                fqName = "com.example.data.UserEntity",
-                packageName = "com.example.data",
-                isInterface = false,
-                isAbstract = false,
-                annotations = listOf(entityAnnotation),
-                imports = emptyList(),
-                referencedTypes = emptySet(),
-                filePath = "/src/UserEntity.kt",
-            )
-        val classWithLeakingConstructor =
-            ClassDeclaration(
-                name = "UserService",
-                fqName = "com.example.UserService",
-                packageName = "com.example",
-                isInterface = false,
-                isAbstract = false,
-                annotations = emptyList(),
-                imports = listOf("com.example.data.UserEntity"),
-                referencedTypes = emptySet(),
-                filePath = "/src/UserService.kt",
-                primaryConstructor =
-                    ConstructorDeclaration(
-                        visibility = Visibility.PUBLIC,
-                        parameters =
-                            listOf(
-                                ParameterDeclaration(
-                                    name = "userEntity",
-                                    type = "UserEntity",
-                                    hasDefaultValue = false,
-                                    annotations = emptyList(),
-                                ),
-                            ),
-                        annotations = emptyList(),
-                    ),
-            )
-
-        val rule =
-            ClassesRuleBuilder(projectGraph)
-                .should()
-                .notHaveSignaturesWithTypesAnnotatedWith("jakarta.persistence.Entity")
-        val assertion = rule.getShouldAssertion()!!
-        val violations = mutableListOf<String>()
-        assertion(classWithLeakingConstructor, listOf(entityClass, classWithLeakingConstructor), violations)
-        assertEquals(1, violations.size)
-        assertTrue(violations[0].contains("UserEntity"))
-        assertTrue(violations[0].contains("Entity"))
     }
 }
