+++ library/src/test/kotlin/io/github/baole/konture/ClassesThatCoverageTest.kt
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-@Suppress("LargeClass")
 internal class ClassesThatCoverageTest : KontureScopeTestFixture() {
     private lateinit var graph: ProjectGraph
